+++ react-components/src/architecture/base/commands/BaseEditTool.ts
@@ -34,10 +34,6 @@ export abstract class BaseEditTool extends NavigationTool {
     return 'crosshair';
   }
 
-  public override setCursor(value: string): void {
-    this.renderTarget.cursor = value;
-  }
-
   public override clearDragging(): void {
     super.clearDragging();
     this._dragger = undefined;
