+++ react-components/src/common/i18n/en/reveal-react-components.json
@@ -129,7 +129,6 @@
   "TOUCH_SELECT": "Tap to select",
   "TOUCH_ZOOM": "Zoom",
   "TITLE": "Title",
-  "TITLE_PLACEHOLDER": "This is the title",
   "UNDO": "Undo",
   "WIDGET_WINDOW_CLOSE": "Close",
   "WIDGET_WINDOW_EXPAND": "Expand",
